libs/utils/analysis: add STATUS analysis module
This patch move the Overutilized analysis code, originally part of the
trace_analysis.py module, to be a new TraceAnalysis module with dynamic
registration via the AnalysisRegister.
